+++ constants/index.js
@@ -1,0 +1,14 @@
+const GOOGLE_SEARCH_LINK = 'http://google.com/search?q=';
+
+const SITE_STACKOVERFLOW = 'site:stackoverflow.com';
+const SITE_YOUTUBE = 'site:youtube.com';
+const SITE_GITHUB = 'site:github.com';
+const SITE_NONE = '';
+
+module.exports = {
+    GOOGLE_SEARCH_LINK,
+    SITE_STACKOVERFLOW,
+    SITE_YOUTUBE,
+    SITE_GITHUB,
+    SITE_NONE
+}