+++ searches/youtubeSearch/index.js
@@ -1,0 +1,10 @@
+const { SITE_YOUTUBE } = require('../../constants');
+const { allErrorsInQuickPick } = require('../')
+
+const searchInYoutube = (currentlyOpenTabfile) => {
+    allErrorsInQuickPick(currentlyOpenTabfile, SITE_YOUTUBE);
+}
+
+module.exports = {
+    searchInYoutube
+}