+++ CHANGELOG.md
@@ -2,8 +2,19 @@
 
 This is a simple extension to search your errors in stackoverflaw. Will add more features soon.
 
-Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+## [0.0.2] - 2021-06-04
+### Added
+- Initial release
+- Search on StackOverflow functionality
+
+## [0.0.3] - 2021-06-05
+### Added
+- Search on Google functionality
+- Search on Github functionality
+- Search on Youtube functionality
+
+### Changed
+- Formatted the Changelog
+- Updated the ReadMe.md
 
 ## [Unreleased]
-
-- Initial release